+++ Models/DoctorSchedule.cs
@@ -7,12 +7,12 @@ namespace VaccineAPI.Models
     public class DoctorSchedule
     {
         public long Id { get; set; }
-        public int DoseId { get; set; }
+        public long DoseId { get; set; }
         
          [JsonIgnore]
         public Dose Dose { get; set; }
        
-        public int DoctorId { get; set; }
+        public long DoctorId { get; set; }
          [JsonIgnore]
         public Doctor Doctor { get; set; }
